+++ puma/tests/test_metrics.py
@@ -8,10 +8,74 @@ import unittest
 
 import numpy as np
 
-from puma.metrics import eff_err, rej_err
+from puma.metrics import calc_separation, eff_err, rej_err
 from puma.utils import logger, set_log_level
 
 set_log_level(logger, "DEBUG")
+
+
+class separation_TestCase(unittest.TestCase):
+    """Test class for the puma.metrics.calc_separation."""
+
+    def setUp(self):
+        """Define a default (seeded) random number generator for all tests"""
+        self.rng = np.random.default_rng(42)
+
+    def test_equal_datasets(self):
+        """Separation of two equal datasets (=0)"""
+        values_a = np.array([1, 1, 2, 2])
+        self.assertEqual(0, calc_separation(values_a, values_a)[0])
+
+    def test_completely_separated(self):
+        """Separation of two completely separated distributions"""
+        values_a = np.array([0.1, 0.5, 0.8, 1])
+        values_b = np.array([1.1, 1.5, 1.8, 2])
+        self.assertAlmostEqual(1, calc_separation(values_a, values_b)[0])
+
+    def test_completely_separated_bad_binning(self):
+        """Separation of two completely separated distributions if the binning
+        if chosen such that they share one bin"""
+        values_a = np.array([0.1, 0.5, 0.8, 1])
+        values_b = np.array([1.1, 1.5, 1.8, 2])
+        self.assertNotAlmostEqual(1, calc_separation(values_a, values_b, bins=3)[0])
+
+    def test_half_separated(self):
+        """Separation of 0.5"""
+        values_a = np.array([0, 1])
+        values_b = np.array([1, 2])
+        self.assertAlmostEqual(0.5, calc_separation(values_a, values_b, bins=3)[0])
+
+    def test_return_bins(self):
+        """Test if bins are correctly returned"""
+        values_a = np.array([0, 1])
+        values_b = np.array([1, 2])
+
+        _, _, hist_a, hist_b, bin_edges = calc_separation(
+            values_a,
+            values_b,
+            bins=3,
+            return_hist=True,
+        )
+        # Check for correct values in hist_a, hist_b and bin_edges
+        np.testing.assert_array_equal(np.array([0.5, 0.5, 0]), hist_a)
+        np.testing.assert_array_equal(np.array([0, 0.5, 0.5]), hist_b)
+        np.testing.assert_array_equal(np.array([0, 2 / 3, 4 / 3, 2]), bin_edges)
+
+    def test_bins_range(self):
+        """Test if bins_range is properly treated."""
+        values_a = np.array([0, 1])
+        values_b = np.array([1, 2])
+
+        _, _, hist_a, hist_b, bin_edges = calc_separation(
+            values_a,
+            values_b,
+            bins=4,
+            bins_range=(0, 4),  # this should result in bin edges [0, 1, 2, 3, 4]
+            return_hist=True,
+        )
+        # Check for correct values in hist_a and bin_edges
+        np.testing.assert_array_equal(np.array([0.5, 0.5, 0, 0]), hist_a)
+        np.testing.assert_array_equal(np.array([0, 1, 2, 3, 4]), bin_edges)
 
 
 class eff_err_TestCase(unittest.TestCase):